+++ src/Application/Features/Roles/RoleValidators.cs
@@ -1,0 +1,21 @@
+using FluentValidation;
+
+namespace Application.Features.Roles;
+
+public sealed class CreateRoleRequestValidator : AbstractValidator<CreateRoleRequest>
+{
+    public CreateRoleRequestValidator()
+    {
+        RuleFor(x => x.Name).NotEmpty().MaximumLength(100);
+        RuleFor(x => x.Description).MaximumLength(400);
+    }
+}
+
+public sealed class UpdateRoleRequestValidator : AbstractValidator<UpdateRoleRequest>
+{
+    public UpdateRoleRequestValidator()
+    {
+        RuleFor(x => x.Name).NotEmpty().MaximumLength(100);
+        RuleFor(x => x.Description).MaximumLength(400);
+    }
+}